Move the messages test to the corresponding file.
